Training Material/packages_logger/requirements.txt
PyMuPDF
Pillow
fastapi 
uvicorn
